fix: 修复 ViolationAnalyzer 和 PlateRecognizer 编译错误
- 修复 minOf 类型错误（Double -> Float）
- 移除 PlateRecognizer 中未使用的 ONNX Runtime 导入

+++ app/src/main/java/com/dva/app/infrastructure/ml/ViolationAnalyzer.kt
@@ -123,7 +123,7 @@ class LaneChangeViolationAnalyzer : ViolationAnalyzer {
         val direction = if (deltaX > 0) "右" else "左"
         
         // 简化置信度计算
-        val confidence = minOf(0.95f, kotlin.math.abs(deltaX) / 100f)
+        val confidence = minOf(0.95f, (kotlin.math.abs(deltaX) / 100.0).toFloat())
         
         return LaneChangeResult(
             confidence = confidence,
